pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "techno-kids"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.32.0",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "bcrypt>=4.0.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.100.0",
    "python-multipart>=0.0.6",
    "httpx>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.0.0",
    "gunicorn>=21.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
